+++ lib/solargraph/rbs_map/core_map.rb
@@ -5,23 +5,36 @@ module Solargraph
     # Ruby core pins
     #
     class CoreMap
-      include Conversions
+
+      FILLS_DIRECTORY = File.join(File.dirname(__FILE__), '..', '..', '..', 'rbs', 'fills')
 
       def initialize
         cache = Cache.load('core.ser')
         if cache
           pins.replace cache
         else
-          loader = RBS::EnvironmentLoader.new(repository: RBS::Repository.new(no_stdlib: false))
-          RBS::Environment.from_loader(loader).resolve_type_names
-          load_environment_to_pins(loader)
-          pins.concat RbsMap::CoreFills::ALL
+          loader.add(path: Pathname(FILLS_DIRECTORY))
+          @pins = conversions.pins
+          @pins.concat RbsMap::CoreFills::ALL
           processed = ApiMap::Store.new(pins).pins.reject { |p| p.is_a?(Solargraph::Pin::Reference::Override) }
-          processed.each { |pin| pin.source = :rbs }
           pins.replace processed
 
           Cache.save('core.ser', pins)
         end
+      end
+
+      def pins
+        @pins ||= []
+      end
+
+      private
+
+      def loader
+        @loader ||= RBS::EnvironmentLoader.new(repository: RBS::Repository.new(no_stdlib: false))
+      end
+
+      def conversions
+        @conversions ||= Conversions.new(loader: loader)
       end
     end
   end