+++ spec/pin/parameter_spec.rb
@@ -154,7 +154,6 @@ describe Solargraph::Pin::Parameter do
     expect(method.documentation).to eq("Block Params:\n*  [Array] \n\nBlock Returns:\n* [Integer] \n\nReturns:\n* [Integer] \n\nVisibility: public")
     expect(method.return_type.tag).to eq('Integer')
 
-
     expect(map.locals.map(&:to_rbs)).to eq(['blk ::Proc', 'things Set'])
     expect(map.locals.map(&:return_type).map(&:to_s)).to eq(%w[Proc Set])
     expect(map.locals.map(&:decl)).to eq(%i[blockarg arg])
@@ -193,21 +192,12 @@ describe Solargraph::Pin::Parameter do
     expect(type.namespace).to eq('Foo::Bar')
   end
 
-  it 'merges near equivalents' do
+  it 'uses longer comment while combining compatible parameters' do
     loc = Solargraph::Location.new('test.rb', Solargraph::Range.from_to(0, 0, 0, 0))
     block = Solargraph::Pin::Block.new(location: loc, name: 'Foo')
     pin1 = Solargraph::Pin::Parameter.new(closure: block, name: 'bar')
     pin2 = Solargraph::Pin::Parameter.new(closure: block, name: 'bar', comments: 'a comment')
-    expect(pin1.try_merge!(pin2)).to be(true)
-  end
-
-  it 'does not merge block parameters from different blocks' do
-    loc = Solargraph::Location.new('test.rb', Solargraph::Range.from_to(0, 0, 0, 0))
-    block1 = Solargraph::Pin::Block.new(location: loc, name: 'Foo')
-    block2 = Solargraph::Pin::Block.new(location: loc, name: 'Bar')
-    pin1 = Solargraph::Pin::Parameter.new(closure: block1, name: 'bar')
-    pin2 = Solargraph::Pin::Parameter.new(closure: block2, name: 'bar', comments: 'a comment')
-    expect(pin1.try_merge!(pin2)).to be(false)
+    expect(pin1.combine_with(pin2).comments).to eq('a comment')
   end
 
   it 'infers undefined types by default' do