+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.55.4 - June 27, 2025
+-  Flatten results of DocMap external bundle query (#981)
+
+## 0.55.3 - June 25, 2025
+-  Nil guards in flow-sensitive typing (patch release) (#980)
+
+## 0.55.2 - June 21, 2025
+- Require external bundle (#972)
+
 ## 0.55.1 - June 8, 2025
 - Fix inline Struct definition (#962)
 - Ensure DocMap requires bundler when loading gemspecs (#963)